Bug fixes
Changed some option names as requested.
Added a few new Aliases.
Changed blocklist to include Idle by default.
Improved sorting in the event of a tie.
Made index=0 by default and adjusted code to suit that.
Added translations (Langauge correction is still a WIP)
Fixed a bug where if a skin were unloaded at the wrong time it would
cause a crash. My fix was adding locks when Rainmeter Adds/Removes a
measure, I also pulled it off the main thread if we need to lock to
prevent a stutter in Rainmeter (Although when it would happen would not
be a big deal)